Fix missing utils package for model_tests
