+++ pipeline/model_tests/chestnut_dec_may/evaluate.py
@@ -8,7 +8,7 @@ from sklearn.metrics import confusion_matrix
 from frdc.train import FRDCDataModule
 from frdc.train import FRDCModule
 from pipeline.model_tests.chestnut_dec_may.preprocess import preprocess
-from pipeline.model_tests.chestnut_dec_may.utils import get_dataset
+from pipeline.model_tests.utils import get_dataset
 
 # Get our Test
 # TODO: Ideally, we should have a separate dataset for testing.
